PLAT-24017 Automatically add links for catalogue widget
This commit modifies the `[administrate-widget]` shortcode so that if
the `product_route` argument is passed to it, this will be used to
specify a link builder function for the Catalogue widget using that
`product_route` and the products name in kebab-case. E.g. if the
`product_route` is `/foo`, Course Bar would be available at
`https://mywebsite.test/foo/course-bar`. If the `configuration` argument
defines the `links` object, then the `product_route` is ignored.