PLAT-24017 Update product route logic based on feedback
The new `product_route` shortcode property that we added has been
updated based on feedback. The two changes are:
1. The route no longer requires a preceding slash, e.g. it was expecting
something like `/courses` but now it just expects `courses`.
2. The function to turn the product name into kebab case has been
changed to mimic the behaviour of Wordpress's `sanitize_title` fucntion.
This is what generates the URL on the server side. We cannot get a
perfect javascript implementation for it, but this should do for most
scenarios.